do not flood logs at warns level about vpnd down
