+++ nym-vpn-app/src-tauri/src/sentry.rs
@@ -1,10 +1,8 @@
-use anyhow::Result;
 use sentry::{ClientInitGuard, User};
 use std::time::Duration;
-use tracing::{error, info, instrument, warn};
+use tracing::{info, instrument, warn};
 
 use crate::env::APP_SENTRY_DSN;
-use crate::grpc::client::GrpcClient;
 use crate::sys::OsInfo;
 
 #[instrument(skip_all)]
@@ -39,29 +37,4 @@ pub fn init(os: &OsInfo) -> Option<ClientInitGuard> {
         }));
     });
     Some(guard)
-}
-
-// Check the state of sentry monitoring at daemon level and
-// sync it with the app state if needed (as shown to the user in UI)
-#[instrument(skip(grpc))]
-pub async fn vpnd_check(sentry_enabled: bool, grpc: &GrpcClient) -> Result<()> {
-    let vpnd_enabled = grpc.sentry_enabled().await.inspect_err(|e| {
-        error!("failed to check sentry state: {:?}", e);
-    })?;
-    if vpnd_enabled == sentry_enabled {
-        // all good
-        return Ok(());
-    }
-    warn!(
-        "sentry state mismatch: app sentry enabled: {}, vpnd sentry enabled: {}",
-        sentry_enabled, vpnd_enabled
-    );
-    if sentry_enabled {
-        info!("enabling vpnd sentry monitoring");
-        grpc.enable_sentry().await?;
-    } else {
-        info!("disabling vpnd sentry monitoring");
-        grpc.disable_sentry().await?;
-    }
-    Ok(())
 }
