+++ nym-vpn-app/src-tauri/src/vpnd.rs
@@ -1,4 +1,4 @@
-use tracing::{error, info, instrument};
+use tracing::{error, info, instrument, warn};
 
 use crate::db::{Db, Key};
 use crate::grpc::client::GrpcClient;
@@ -28,6 +28,31 @@ pub async fn netstats_check(db: &Db, grpc: &GrpcClient) -> anyhow::Result<()> {
     } else {
         info!("disabled vpnd network statistics collection");
         grpc.disable_netstats().await?;
+    }
+    Ok(())
+}
+
+// Check the state of sentry monitoring in daemon side
+// if needed sync it with the saved setting from the app db
+#[instrument(skip(grpc))]
+pub async fn sentry_check(sentry_enabled: bool, grpc: &GrpcClient) -> anyhow::Result<()> {
+    let vpnd_enabled = grpc.sentry_enabled().await.inspect_err(|e| {
+        error!("failed to check sentry state: {:?}", e);
+    })?;
+    if vpnd_enabled == sentry_enabled {
+        // all good
+        return Ok(());
+    }
+    warn!(
+        "sentry state mismatch: app sentry enabled: {}, vpnd sentry enabled: {}",
+        sentry_enabled, vpnd_enabled
+    );
+    if sentry_enabled {
+        info!("enabling vpnd sentry monitoring");
+        grpc.enable_sentry().await?;
+    } else {
+        info!("disabling vpnd sentry monitoring");
+        grpc.disable_sentry().await?;
     }
     Ok(())
 }
